Remove the CanOverwriteFiles reference.
Because we are wrapping Filesystems and not merely adapters, implementing CanOverwriteFiles means that we run into an issue where a put() operation can simply write() to a file but when the assertion occurs at the actualy Adapter level, the write() contains yet-another assertion to see if the file exists. At this point it is too late to catch it and so our only workaround at this point is to just have it do the assertion and use update instead.