Avoid out of memory issues for large files.
Not entirely sure where I copied this from originally, but it has been causing issues with allocating far too much memory for files. Rather than load the file completely and then get the length of the string, just have PHP seek to the end of the file and then figure out where the pointer is. This seems to not only be much faster but also stops OOM issues with files that exceed the memory_limit.