Use bind instead of singleton.
In some cases, it is possible for metadata failures to occur because the shared singleton is the wrong thing to use. Each disk requires it's own config. If the "storage ID" is different and you create two disk instances, you can end up in a situation where shared singleton now points to the wrong metadata storage ID and stuff begins to fail...